requirements.txt
streamlit
pandas
plotly
transformers
torch
scikit-learn
python-dotenv
groq
